% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{colleziona}
\alias{colleziona}
\title{Collect article links from a single page}
\usage{
colleziona(x, autlink)
}
\arguments{
\item{x}{URL of a page to scrape}

\item{autlink}{Author archive base URL used to identify article position}
}
\value{
Character vector of article URLs
}
\description{
Extracts article links from a paginated author archive page.
}
\keyword{internal}
