% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mieitesti-package.R
\docType{package}
\name{mieitesti-package}
\alias{mieitesti-package}
\alias{mieitesti}
\title{mieitesti: Italian News Article Collection and Text Analysis}
\description{
Tools for collecting and analyzing Italian news articles from
'il sussidiario' and ISTAT RSS feeds.
}
\section{Core Functions}{

\itemize{
\item \code{\link{sussidiario}} - Collect all articles from an author
\item \code{\link{leggi_pagina}} - Parse a single article page
\item \code{\link{pag_as_frame}} - Convert article list to data.table
\item \code{\link{sintesi}} - Extract ISTAT article summaries
}
}

\section{Text Utilities}{

\itemize{
\item \code{\link{apostrofo}} - Normalize Italian apostrophes
}
}

\section{Data}{

\itemize{
\item \code{\link{miestop}} - Italian stopwords character vector
\item \code{\link{emp_dictionary}} - Employment topic dictionary
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gmontaletti/mieitesti}
  \item Report bugs at \url{https://github.com/gmontaletti/mieitesti/issues}
}

}
\author{
Giampaolo Montaletti \email{giampaolo.montaletti@gmail.com}
}
