% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{sussidiario}
\alias{sussidiario}
\title{Collect all articles from an author on il sussidiario}
\usage{
sussidiario(autore)
}
\arguments{
\item{autore}{URL of the author's archive page on ilsussidiario.net}
}
\value{
Character vector of article URLs
}
\description{
Scrapes the ilsussidiario.net author archive to collect all article URLs.
This function handles pagination automatically.
}
\examples{
\dontrun{
autlink <- "https://www.ilsussidiario.net/autori/giampaolo-montaletti/"
articoli <- sussidiario(autore = autlink)
}
}
